Refactored the code base and created the following:
LauncherActivity.java as the main point of the application where it will determine if the user logged in or direct to the proper page.

SessionManager.java is a util class that makes handling the user session easier.

This commit also include the functionality for signing in and signing up for a new account along with the ability to logout.